Refactor Security Folder
- Se movió la carpeta de Security a Core/Infrastructure
