+++ src/core/infrastructure/security/auth/sessions/exceptions/invalid.session.exception.ts
