+++ src/core/infrastructure/security/auth/sessions/session.guard.ts
