+++ src/core/infrastructure/security/auth/sessions/session.serializer.ts
