+++ src/core/infrastructure/security/users/decorators/get-doctor-id.decortator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from "@nestjs/common"
-import { InvalidSessionException } from "src/security/auth/sessions/exceptions/invalid.session.exception";
+import { InvalidSessionException } from "../../auth/sessions/exceptions/invalid.session.exception";
 
 export const GetDoctorId = createParamDecorator(
     (_data, context: ExecutionContext) => {
