+++ src/core/infrastructure/security/users/entities/user.entity.ts
@@ -1,6 +1,6 @@
+import { OrmDoctor } from "src/doctor/infrastructure/entities/orm-doctor.entity";
+import { OrmPatient } from "src/patient/infrastructure/entities/orm-patient.entity";
 import { Check, Column, CreateDateColumn, Entity, Index, JoinColumn, OneToMany, OneToOne, PrimaryGeneratedColumn, UpdateDateColumn } from "typeorm";
-import { OrmDoctor } from "../../../doctor/infrastructure/entities/orm-doctor.entity";
-import { OrmPatient } from "../../../patient/infrastructure/entities/orm-patient.entity";
 import { SessionEntity } from "../../../security/auth/sessions/entities/session.entity";
 import { Role } from "../roles/role.entity.enum";
