+++ src/core/infrastructure/security/users/roles/role.entity.enum.ts
